Refactor progress components to use a new Progress class for better code reuse
