+++ resources/views/tables/columns/circle-progress.blade.php
@@ -1,28 +1,14 @@
 @php
-    $total = $getState()['total'];
-    $progress = $getState()['progress'];
-    $progress = $total > 0 ? ($progress / $total) * 100 : 0;
-
-    if($progress === 100){
-        $progressColor = '#2980b9';
-    } else if($progress > 50){
-        $progressColor = '#27ae60';
-    } else if($progress > 25){
-        $progressColor = '#f39c12';
-    } else {
-        $progressColor = '#e74c3c';
-    }    
-    
-    $displayProgress = $progress === 100 ? number_format($progress, 0) : number_format($progress, 2);
+    $data = $column->getProgressData($getState());
 @endphp
 
 <div class="progress-circle" style="
     background: conic-gradient(
-        {{ $progressColor }} {{ $displayProgress * 3.6 }}deg,
-        #e5e7eb {{ $displayProgress * 3.6 }}deg
+        {{ $data['progressColor'] }} {{ $data['displayProgress'] * 3.6 }}deg,
+        #e5e7eb {{ $data['displayProgress'] * 3.6 }}deg
     );">
     @if($column instanceof \LaraZeus\Progress\Tables\Columns\CircleProgress && $column->getCanShow())
-        <small>{{ $displayProgress }}%</small>
+        <small>{{ $data['displayProgress'] }}%</small>
     @endif
 </div>
