+++ src/Tables/Columns/CircleProgress.php
@@ -2,23 +2,7 @@
 
 namespace LaraZeus\Progress\Tables\Columns;
 
-use Filament\Tables\Columns\Column;
-
-class CircleProgress extends Column
+class CircleProgress extends Progress
 {
-    protected bool $canShow = true;
-
     protected string $view = 'lara-zeus-progress::tables.columns.circle-progress';
-
-    public function hideProgressValue($canShow = false): static
-    {
-        $this->canShow = $canShow;
-
-        return $this;
-    }
-
-    public function getCanShow(): bool
-    {
-        return $this->canShow;
-    }
 }
