+++ src/Tables/Columns/ProgressBar.php
@@ -2,23 +2,7 @@
 
 namespace LaraZeus\Progress\Tables\Columns;
 
-use Filament\Tables\Columns\Column;
-
-class ProgressBar extends Column
+class ProgressBar extends Progress
 {
-    protected bool $canShow = true;
-
     protected string $view = 'lara-zeus-progress::tables.columns.progress-bar';
-
-    public function hideProgressValue($canShow = false): static
-    {
-        $this->canShow = $canShow;
-
-        return $this;
-    }
-
-    public function getCanShow(): bool
-    {
-        return $this->canShow;
-    }
 }
